+++ packages/gamut/src/Form/Select.tsx
@@ -5,25 +5,31 @@ import {
 import { variant } from '@codecademy/gamut-styles';
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
-import { each, isArray, isObject } from 'lodash';
 import React, {
   ChangeEvent,
   forwardRef,
-  ReactNode,
   SelectHTMLAttributes,
+  useMemo,
   useState,
 } from 'react';
 
 import { Box, FlexBox } from '../Box';
 import { conditionalStyles, formFieldStyles } from './styles/shared';
+import { parseSelectOptions } from './utils';
 
-export type SelectWrapperProps = SelectHTMLAttributes<HTMLSelectElement> & {
+export type SelectComponentProps = Pick<
+  SelectHTMLAttributes<HTMLSelectElement>,
+  'disabled' | 'id'
+> & {
   error?: boolean;
   htmlFor?: string;
   options?: string[] | Record<string, number | string>;
-  id?: string;
-  sizeVariant?: 'small' | 'base';
 };
+
+export type SelectWrapperProps = SelectComponentProps &
+  SelectHTMLAttributes<HTMLSelectElement> & {
+    sizeVariant?: 'small' | 'base';
+  };
 
 export interface SelectProps extends SelectWrapperProps {
   activated?: boolean;
@@ -40,6 +46,7 @@ const selectSizeVariants = variant({
     },
     base: {
       height: 'auto',
+      paddingRight: 48,
     },
   },
 });
@@ -55,12 +62,11 @@ const SelectBase = styled.select<SelectProps>`
   appearance: none;
 `;
 
-const selectIconStyles = css`
+const allowClickStyle = css`
   pointer-events: none;
 `;
 
-const SelectIcon = styled(ArrowChevronDownIcon)(selectIconStyles);
-const MiniSelectIcon = styled(MiniChevronDownIcon)(selectIconStyles);
+const StyledFlexbox = styled(FlexBox)(allowClickStyle);
 
 export const Select = forwardRef<HTMLSelectElement, SelectWrapperProps>(
   (
@@ -74,27 +80,9 @@ export const Select = forwardRef<HTMLSelectElement, SelectWrapperProps>(
       setActivated(true);
     };
 
-    let selectOptions: ReactNode[] = [];
-
-    if (isArray(options)) {
-      selectOptions = options.map((option) => {
-        const key = id ? `${id}-${option}` : option;
-        return (
-          <option key={key} value={option} data-testid={key}>
-            {option}
-          </option>
-        );
-      });
-    } else if (isObject(options)) {
-      each(options, (text, val) => {
-        const key = id ? `${id}-${val}` : val;
-        selectOptions.push(
-          <option key={key} value={val} data-testid={key}>
-            {text}
-          </option>
-        );
-      });
-    }
+    const selectOptions = useMemo(() => {
+      return parseSelectOptions({ options, id });
+    }, [options, id]);
 
     return (
       <Box
@@ -104,20 +92,21 @@ export const Select = forwardRef<HTMLSelectElement, SelectWrapperProps>(
         minWidth="7rem"
         className={className}
       >
-        <FlexBox
-          paddingRight={12}
+        <StyledFlexbox
+          pr={12}
           alignItems="center"
           position="absolute"
           right="0"
           top="0"
           bottom="0"
+          aria-hidden
         >
           {sizeVariant === 'small' ? (
-            <MiniSelectIcon size={12} />
+            <MiniChevronDownIcon size={12} />
           ) : (
-            <SelectIcon size={16} />
+            <ArrowChevronDownIcon size={16} />
           )}
-        </FlexBox>
+        </StyledFlexbox>
         <SelectBase
           {...rest}
           defaultValue={defaultValue || ''}