+++ packages/gamut/src/deprecated/Icon/defaultIconProps.tsx
@@ -1,8 +1,0 @@
-export const defaultIconProps = {
-  fill: 'currentColor',
-  height: 24,
-  version: '1.1',
-  viewBox: '0 0 24 24',
-  width: 24,
-  xmlns: 'http://www.w3.org/2000/svg',
-};