+++ packages/styleguide/.storybook/main.ts
@@ -42,6 +42,10 @@ module.exports = {
       alias: {
         ...emotionless(config.resolve.alias),
         '~styleguide/blocks': path.resolve(__dirname, './components/'),
+        '@codecademy/storybook-addon-variance': path.resolve(
+          __dirname,
+          './addons/system/components/'
+        ),
         '@codecademy/gamut-styles$': path.resolve(
           __dirname,
           '../../gamut-styles/src'
@@ -58,6 +62,10 @@ module.exports = {
         '@codecademy/gamut-illustrations$': path.resolve(
           __dirname,
           '../../gamut-illustrations/src'
+        ),
+        '@codecademy/gamut-patterns$': path.resolve(
+          __dirname,
+          '../../gamut-patterns/src'
         ),
       },
     };