+++ packages/gamut-styles/CHANGELOG.md
@@ -3,6 +3,126 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.8.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@9.7.4...@codecademy/gamut-styles@9.8.0) (2021-05-05)
+
+
+### Features
+
+* **PageWrapper:** Extend stacking context reset element for pages with a background color ([0a22b34](https://github.com/Codecademy/client-modules/commit/0a22b3489625e524085ffd2076730e05b9a4d898))
+
+
+
+### [9.7.4](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@9.7.3...@codecademy/gamut-styles@9.7.4) (2021-05-04)
+
+**Note:** Version bump only for package @codecademy/gamut-styles
+
+
+
+
+
+### [9.7.3](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@9.7.2...@codecademy/gamut-styles@9.7.3) (2021-05-04)
+
+
+### Bug Fixes
+
+* **Props/Dialog/Button:** Prop Fixes for Button / Dialog ([9b03d03](https://github.com/Codecademy/client-modules/commit/9b03d0336f545f5107ba600a953f6f04edcf97da))
+
+
+
+### [9.7.2](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@9.7.1...@codecademy/gamut-styles@9.7.2) (2021-05-03)
+
+**Note:** Version bump only for package @codecademy/gamut-styles
+
+
+
+
+
+### [9.7.1](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@9.7.0...@codecademy/gamut-styles@9.7.1) (2021-04-30)
+
+
+### Bug Fixes
+
+* **Emotion:** Add mode and variant to base `shouldForwardProp` filters and keep static list of all system props ([19b34e4](https://github.com/Codecademy/client-modules/commit/19b34e40abf9bf1ec2a92f4828c734ec67ee879e))
+
+
+
+## [9.7.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@9.6.3...@codecademy/gamut-styles@9.7.0) (2021-04-29)
+
+
+### Features
+
+* **Buttons:** ColorModes + Props + Refactor + Refs Oh My! [GM-211] [REACH-856] [GM-192] ([c4643db](https://github.com/Codecademy/client-modules/commit/c4643db64d3684319f8d5fc6f68d05197ec1f61e))
+
+
+
+### [9.6.3](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@9.6.2...@codecademy/gamut-styles@9.6.3) (2021-04-28)
+
+**Note:** Version bump only for package @codecademy/gamut-styles
+
+
+
+
+
+### [9.6.2](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@9.6.1...@codecademy/gamut-styles@9.6.2) (2021-04-27)
+
+**Note:** Version bump only for package @codecademy/gamut-styles
+
+
+
+
+
+### [9.6.1](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@9.6.0...@codecademy/gamut-styles@9.6.1) (2021-04-26)
+
+
+### Bug Fixes
+
+* **ColorMode:** Ref Forarding for BG + ColorMode ([d2ded46](https://github.com/Codecademy/client-modules/commit/d2ded46456df895e0bfa58ac213c583c2ff934c1))
+
+
+
+## [9.6.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@9.5.6...@codecademy/gamut-styles@9.6.0) (2021-04-26)
+
+
+### Features
+
+* **Anchor:** Variance Migration + ColorMode [GM-203] ([d5ae675](https://github.com/Codecademy/client-modules/commit/d5ae6758c65a837066562eb0ed8b964176889148))
+
+
+
+### [9.5.6](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@9.5.5...@codecademy/gamut-styles@9.5.6) (2021-04-23)
+
+
+### Bug Fixes
+
+* cache constants missing ([de8094b](https://github.com/Codecademy/client-modules/commit/de8094bd0b21c671f75ae8b6a6b895260b66c2de))
+
+
+
+### [9.5.5](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@9.5.4...@codecademy/gamut-styles@9.5.5) (2021-04-23)
+
+
+### Bug Fixes
+
+* Add GamutProvider Tests and Remove .storybook from coverage ([87bbb73](https://github.com/Codecademy/client-modules/commit/87bbb737a0cd2a9f5430c65b44cd82071b186668))
+
+
+
+### [9.5.4](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@9.5.3...@codecademy/gamut-styles@9.5.4) (2021-04-22)
+
+**Note:** Version bump only for package @codecademy/gamut-styles
+
+
+
+
+
+### [9.5.3](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@9.5.2...@codecademy/gamut-styles@9.5.3) (2021-04-20)
+
+**Note:** Version bump only for package @codecademy/gamut-styles
+
+
+
+
+
 ### [9.5.2](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@9.5.1...@codecademy/gamut-styles@9.5.2) (2021-04-19)