+++ packages/gamut-tests/CHANGELOG.md
@@ -3,6 +3,118 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [2.3.27](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-tests@2.3.26...@codecademy/gamut-tests@2.3.27) (2021-05-05)
+
+**Note:** Version bump only for package @codecademy/gamut-tests
+
+
+
+
+
+### [2.3.26](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-tests@2.3.25...@codecademy/gamut-tests@2.3.26) (2021-05-04)
+
+**Note:** Version bump only for package @codecademy/gamut-tests
+
+
+
+
+
+### [2.3.25](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-tests@2.3.24...@codecademy/gamut-tests@2.3.25) (2021-05-04)
+
+**Note:** Version bump only for package @codecademy/gamut-tests
+
+
+
+
+
+### [2.3.24](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-tests@2.3.23...@codecademy/gamut-tests@2.3.24) (2021-05-03)
+
+**Note:** Version bump only for package @codecademy/gamut-tests
+
+
+
+
+
+### [2.3.23](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-tests@2.3.22...@codecademy/gamut-tests@2.3.23) (2021-04-30)
+
+**Note:** Version bump only for package @codecademy/gamut-tests
+
+
+
+
+
+### [2.3.22](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-tests@2.3.21...@codecademy/gamut-tests@2.3.22) (2021-04-29)
+
+**Note:** Version bump only for package @codecademy/gamut-tests
+
+
+
+
+
+### [2.3.21](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-tests@2.3.20...@codecademy/gamut-tests@2.3.21) (2021-04-28)
+
+**Note:** Version bump only for package @codecademy/gamut-tests
+
+
+
+
+
+### [2.3.20](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-tests@2.3.19...@codecademy/gamut-tests@2.3.20) (2021-04-27)
+
+**Note:** Version bump only for package @codecademy/gamut-tests
+
+
+
+
+
+### [2.3.19](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-tests@2.3.18...@codecademy/gamut-tests@2.3.19) (2021-04-26)
+
+**Note:** Version bump only for package @codecademy/gamut-tests
+
+
+
+
+
+### [2.3.18](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-tests@2.3.17...@codecademy/gamut-tests@2.3.18) (2021-04-26)
+
+**Note:** Version bump only for package @codecademy/gamut-tests
+
+
+
+
+
+### [2.3.17](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-tests@2.3.16...@codecademy/gamut-tests@2.3.17) (2021-04-23)
+
+**Note:** Version bump only for package @codecademy/gamut-tests
+
+
+
+
+
+### [2.3.16](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-tests@2.3.15...@codecademy/gamut-tests@2.3.16) (2021-04-23)
+
+**Note:** Version bump only for package @codecademy/gamut-tests
+
+
+
+
+
+### [2.3.15](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-tests@2.3.14...@codecademy/gamut-tests@2.3.15) (2021-04-22)
+
+**Note:** Version bump only for package @codecademy/gamut-tests
+
+
+
+
+
+### [2.3.14](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-tests@2.3.13...@codecademy/gamut-tests@2.3.14) (2021-04-20)
+
+**Note:** Version bump only for package @codecademy/gamut-tests
+
+
+
+
+
 ### [2.3.13](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-tests@2.3.12...@codecademy/gamut-tests@2.3.13) (2021-04-19)
 
 **Note:** Version bump only for package @codecademy/gamut-tests